+++ rust/opendp/src/trans/sum/mod.rs
@@ -7,7 +7,7 @@ use crate::core::{Function, StabilityRelation, Transformation};
 use crate::dist::{AbsoluteDistance, IntDistance, SymmetricDistance};
 use crate::dom::{AllDomain, BoundedDomain, SizedDomain, VectorDomain};
 use crate::error::*;
-use crate::traits::{Abs, DistanceConstant, InfCast, SaturatingAdd, CheckedMul, ExactIntCast, CheckNull};
+use crate::traits::{Abs, DistanceConstant, InfCast, SaturatingAdd, CheckedMul, ExactIntCast, CheckNull, InfDiv};
 
 pub fn make_bounded_sum<T>(
     bounds: (T, T)
@@ -25,43 +25,6 @@ pub fn make_bounded_sum<T>(
         StabilityRelation::new_from_constant(lower.abs().total_max(upper.abs())?)))
 }
 
-// division with rounding towards infinity
-pub trait InfDiv {
-    fn inf_div(&self, other: &Self) -> Self;
-}
-
-macro_rules! impl_int_inf_div {
-    ($($ty:ty),+) => ($(impl InfDiv for $ty {
-        fn inf_div(&self, other: &Self) -> Self {
-            (self + 1) / other
-        }
-    })+)
-}
-impl_int_inf_div!(u8, u16, u32, u64, u128, i8, i16, i32, i64, i128);
-
-macro_rules! impl_float_inf_div {
-    ($($ty:ty),+) => ($(impl InfDiv for $ty {
-        fn inf_div(&self, other: &Self) -> Self {
-            let div = self / other;
-            if !div.is_finite() {
-                // don't increment -Inf or Inf into a NaN, leave NaN as-is
-                div
-            } else if div * other <= *self {
-                // < is (probably) too tight, <= is too loose. Remain conservative with <=
-                // perturb the floating-point bit representation by taking the next float
-                <$ty>::from_bits(if div.is_sign_negative() {
-                    div.to_bits() - 1
-                } else {
-                    div.to_bits() + 1
-                })
-            } else {
-                div
-            }
-        }
-    })+)
-}
-impl_float_inf_div!(f32, f64);
-
 pub fn make_sized_bounded_sum<T>(
     size: usize, bounds: (T, T)
 ) -> Fallible<Transformation<SizedDomain<VectorDomain<BoundedDomain<T>>>, AllDomain<T>, SymmetricDistance, AbsoluteDistance<T>>>
@@ -74,7 +37,6 @@ pub fn make_sized_bounded_sum<T>(
         return fallible!(MakeTransformation, "Detected potential for overflow when computing function.")
     }
     let _2 = T::exact_int_cast(2)?;
-    let range = upper - lower;
     Ok(Transformation::new(
         SizedDomain::new(VectorDomain::new(
             BoundedDomain::new_closed(bounds)?), size),
@@ -82,15 +44,9 @@ pub fn make_sized_bounded_sum<T>(
         Function::new(|arg: &Vec<T>| arg.iter().sum()),
         SymmetricDistance::default(),
         AbsoluteDistance::default(),
-        // naively:
-        // d_out >= d_in * (M - m) / 2
-        // to avoid integer truncation:
-        // d_out * 2 >= d_in * (M - m)
-        StabilityRelation::new_all(
-            enclose!((_2, range), move |&d_in: &IntDistance, d_out: &T|
-                Ok(d_out.clone() * _2.clone() >= T::inf_cast(d_in)? * range.clone())),
-            Some(move |d_in: &IntDistance| Ok(Box::new((T::inf_cast(*d_in)? * range.clone()).inf_div(&_2)))),
-        None::<fn(&_) -> _>)))
+        // d_out >= d_in * (U - L).inf_div(2)
+        StabilityRelation::new_from_constant(
+            (upper - lower).inf_div(&T::exact_int_cast(2)?))))
 }
